requirements.txt
argcomplete==1.12.3
asgiref==3.4.1
boto3==1.17.111
botocore==1.20.111
certifi==2021.5.30
cffi==1.14.6
cfn-flip==1.2.3
charset-normalizer==2.0.1
click==8.0.1
cryptography==3.4.7
defusedxml==0.7.1
dj-rest-auth==2.1.9
Django==3.2.5
django-allauth==0.44.0
django-cors-headers==3.7.0
djangorestframework==3.12.4
durationpy==0.5
future==0.18.2
hjson==3.0.2
idna==3.2
jmespath==0.10.0
kappa==0.6.0
oauthlib==3.1.1
pep517==0.10.0
pip-tools==6.2.0
placebo==0.9.0
psycopg2-binary==2.9.1
pycparser==2.20
PyJWT==2.1.0
python-dateutil==2.8.2
python-slugify==5.0.2
python3-openid==3.2.0
pytz==2021.1
PyYAML==5.4.1
requests==2.26.0
requests-oauthlib==1.3.0
s3transfer==0.4.2
six==1.16.0
sqlparse==0.4.1
text-unidecode==1.3
toml==0.10.2
tqdm==4.61.2
troposphere==2.7.0
urllib3==1.26.6
Werkzeug==0.16.1
whitenoise==5.2.0
wsgi-request-logger==0.4.6
zappa==0.53.0
